Fixed bug with creation and data checking of very large single file torrents, beats me why these casts are need, you would expect the compiler to figure it out.
svn path=/trunk/extragear/network/ktorrent/; revision=626595

+++ apps/ktcachecheck/singlecachechecker.cpp
@@ -63,7 +63,7 @@ namespace ktdebug
 			{
 				Uint32 size = i == num_chunks - 1 && tor.getFileLength() % tor.getChunkSize() > 0 ?
 						tor.getFileLength() % tor.getChunkSize() : (Uint32)tor.getChunkSize();
-				fptr.seek(File::BEGIN,i*tor.getChunkSize());
+				fptr.seek(File::BEGIN,(Int64)i*tor.getChunkSize());
 				fptr.read(buf,size);
 				SHA1Hash h = SHA1Hash::generate(buf,size);
 				bool ok = (h == tor.getHash(i));
